mcp/pyproject.toml
[project]
name = "memocc"
version = "0.2.0"
description = "FastMCP server for engram - a portable memory graph for Claude Code interactions"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.3",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "sqlmodel>=0.0.22",
    "alembic>=1.14.0",
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.6.0",
    "psycopg2-binary>=2.9.0",
    "greenlet>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-docker>=3.1.0",
    "vcrpy>=6.0.0",
    "pytest-recording>=0.13.0",
    "factory-boy>=3.3.0",
    "faker>=28.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["app"]
omit = ["app/db/migrations/*"]
